Change project structure. Add separate readme files
