--- .gitattributes
@@ -1,1 +1,0 @@
-*.h5 filter=lfs diff=lfs merge=lfs -text
