+++ GUI_APPLICATION.md
@@ -1,0 +1,9 @@
+# Application
+
+Because it's awful idea to make a project that will work only from command line, GUI application was developed.
+
+<img src="doc/main-window.png" alt="" width="300">
+
+<img src="doc/ndvi-visualization.png" alt="" width="600">
+
+<img src="doc/ndvi-classes-visualization.png" alt="" width="600">
