+++ forestryaccounting/__main__.py
@@ -1,0 +1,4 @@
+from .main import MainGUIController
+
+if __name__ == "__main__":
+    MainGUIController.main()
