+++ scripts/deforestation_reforestation_detection/imageprocessing.py
@@ -1,0 +1,1 @@
+../../forestryaccounting/imageprocessing.py
